backend/scene-detector/pyproject.toml
[project]
name = "scene-detector"
version = "0.1.0"
description = "service that takes in raw videos, splits it into scenes, and sends it downstream for parallel processing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nats-py>=2.14.0",
    "pydantic-settings>=2.13.1",
    "scenedetect[opencv-headless]>=0.6.7.1",
    "structlog>=25.5.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.13"
include = ["src", "test"]
extraPaths = ["."]

[tool.pyrefly]
python-version = "3.13"
untyped-def-behavior = "check-and-infer-return-any"
search-path = ["."]

[tool.pyrefly.errors]
unannotated-parameter = "error"
unannotated-return = "error"
unannotated-attribute = "error"
implicit-any = "error"

[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning:testcontainers",
]

[tool.ruff]
exclude = [".venv"]

[dependency-groups]
dev = [
    "pyrefly>=0.57.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
    "testcontainers[nats]>=4.14.2",
]
